Update Slack API event handling
